feat(server): move permissions sync
