+++ packages/twenty-server/src/engine/core-modules/application/utils/from-field-manifest-to-universal-flat-field-metadata.util.ts
@@ -2,7 +2,7 @@ import {
   type FieldManifest,
   type RelationFieldManifest,
 } from 'twenty-shared/application';
-import { FieldMetadataType } from 'twenty-shared/types';
+import { type FieldMetadataType } from 'twenty-shared/types';
 
 import { generateDefaultValue } from 'src/engine/metadata-modules/field-metadata/utils/generate-default-value';
 import { isMorphOrRelationFieldMetadataType } from 'src/engine/utils/is-morph-or-relation-field-metadata-type.util';
